pyproject.toml
[tool.poetry]
name = "tc_mailmanager"
version = "1.6.0"
description = "ToucanToco's cross-python-projects MailManager"
authors = ["Toucan Toco <dev@toucantoco.com>"]
readme = "README.md"
homepage = "https://github.com/ToucanToco/tc_mailmanager"
documentation = "https://toucantoco.github.io/tc_mailmanager"
repository = "https://github.com/ToucanToco/tc_mailmanager"
license = "BSD-3-Clause"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.8"
sendgrid = "^6.9.6"
tctc-envelopes = "0.5"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.931"
pre-commit = "^2.17.0"
pytest = "^7.0.0"
pytest-cov = "^3.0.0"
pytest-mock = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py310"]
line-length = 100

[tool.isort]
profile = "black"
combine_as_imports = true

[tool.mypy]
python_version = "3.10"
pretty = true
allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
strict_equality = true
no_implicit_optional = true
warn_no_return = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_return_any = true
warn_unreachable = true
show_error_codes = true
show_column_numbers = true
show_error_context = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
